requirements.txt
amqp==2.3.2
astroid==2.0.2
Babel==2.6.0
billiard==3.5.0.4
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0
django-bootstrap3==10.0.1
django-phonenumber-field==2.0.0
djangorestframework==3.7.3
idna==2.6
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mpesa-api==0.1.8
phonenumbers==8.9.10
psycopg2==2.7.5
pylint==2.1.0
pytz==2018.5
requests==2.18.4
six==1.11.0
typed-ast==1.1.0
typing==3.6.4
urllib3==1.22
vine==1.1.4
wrapt==1.10.11
